Rename files to fit new generic functionality
